Move inline tailwind to style tags
Substantial refactor to remove all tailwind class annotations from
markup and into postcss @apply statements on in style tags.
This is a workaround to enable bundling tailwind styles, since the
tailwind preprocessor step does not run on `sveltekit package`.
The alternative would be to "properly" bundle the package (TODO:
investigate) with vite (maybe dropping into rollup?). A tradeoff of this
approach is that the tailwind css normalization (`@tailwind base;`
etc in app.css) is not applied. That is fine so long as the component is
being consumed by another app using tailwind, but will result in
different styles (and probably looking bad) if that is not the case.

Incidentally, removed the papaya background colour, which was applied
with a global. The background colour can be the responsibility of the
consuming app.